backend/requirements.txt
django-admin-sortable==2.1
django-cors-headers==2.0.2
django-debug-toolbar==1.7
django-extensions==1.7.8
django-hashid-field==1.2.1
Django==1.11
djangorestframework==3.6.2
envparse==0.2.0
Pillow==4.1.0  # required for ImageField
psycopg2==2.7.1
requests==2.13.0
sparkpost==1.3.5